Home Project based on clases
This project implements a matrix that has polynomial functions for as
elements and each polynomial has complex numbers as coefficients.
The scope was to create a complex number class and overload its
operators same for polynomial class and matrix class.
In the main file is an example on how this classes can work together.